1.12.2

+++ src/ProtonDrive.Client/Instrumentation/Observability/ObservabilityMetricProperties.cs
@@ -1,3 +1,3 @@
 ﻿namespace ProtonDrive.Client.Instrumentation.Observability;
 
-public sealed record ObservabilityMetricProperties(int Value, IReadOnlyDictionary<string, string> Labels);
+public sealed record ObservabilityMetricProperties(long Value, IReadOnlyDictionary<string, string> Labels);
